// stonks/src/yahoo/query2YahooFinanceV8/interfaces.ts
export interface CurrentTradingPeriod {
  pre: Pre;
  regular: Regular;
  post: Post;
}
export interface Meta {
  currency: string;
  symbol: string;
  exchangeName: string;
  instrumentType: string;
  firstTradeDate: number;
  regularMarketTime: number;
  gmtoffset: number;
  timezone: string;
  exchangeTimezoneName: string;
  regularMarketPrice: number;
  chartPreviousClose: number;
  priceHint: number;
  currentTradingPeriod: CurrentTradingPeriod;
  dataGranularity: string;
  range: string;
  validRanges: string[];
}
export interface Quote {
  close: number[];
  high: number[];
  open: number[];
  volume: number[];
  low: number[];
}
export interface Adjclose {
  adjclose: number[];
}
export interface Indicators {
  quote: Quote[];
  adjclose: Adjclose[];
}
export interface Result {
  meta: Meta;
  timestamp: number[];
  indicators: Indicators;
}
export interface Chart {
  result: Result[];
  error?: any;
}
export interface Query2YahooFinanceV8ChartResponse {
  chart: Chart;
}